lightning/views.py
import os
import tempfile
import yt_dlp as youtube_dl
import ffmpeg
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status
from whisper import load_model

class Transcribe(APIView):
    def get(self, request, *args, **kwargs):
        video_url = request.query_params.get('url')
        
        if not video_url:
            return Response({'error': 'No video URL provided'}, status=status.HTTP_400_BAD_REQUEST)
        try:
            # your logic here
            # steps:
            #     get transcription from the video url (upto 1000 chars)
            #     use scrapy and scrape chatGPT-4o mini (without login) to get the SEO output using the transcript(seo output format is given in the result object)
            #     return the SEO output
            
            result = {
                "title":"",
                "description": "",
                "thumbnailSuggestion":""
            }
            return Response({'SEO output': result})
        
        except Exception as e:
            print("Error:", e)
            return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
