requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.1
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
drf-social-oauth2==3.1.0
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.2.0
future==1.0.0
idna==3.8
Jinja2==3.1.3
jwcrypto==1.5.6
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
oauthlib==3.2.2
openai-whisper==20231117
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytube2==15.0.6
pytz==2024.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
sympy==1.12
tiktoken==0.7.0
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2